Move api code to flaskr directory
