--- API/testdata/getParsedDataTest.json
@@ -1,10 +1,0 @@
-[
-{
-    "title": "M 4.9 - Near Moodus, Connecticut",
-    "mag": 4.9,
-    "time": -12685939200000,
-    "lat": 41.5,
-    "lon": -72.5,
-    "url": "https://earthquake.usgs.gov/fdsnws/event/1/query?eventid=ushis15680101000000000&format=geojson"
-}
-]
