+++ flaskr/test/readme.md
@@ -1,0 +1,3 @@
+# To run tests
+
+Use command `python3 -m unittest -v flaskr.test.APItest`
